+++ lib/app/features/gpu/providers/gpu_providers.dart
@@ -1,6 +1,6 @@
 abstract class GpuProviders {
   Future<String> getGpuName();
   Future<String> getGpuCurrentFrequencyNode();
-  Future<List<int>> getGpuAvailableFrequencies();
+  Future<List<double>> getGpuAvailableFrequencies();
   Future<List<String>> getGpuAvailableGovernors();
 }
